Improve platform USB interop and add structure tests
Add PlatformInteropStructureTests to validate native struct sizes/fields across backends. Update project file to expose internals to tests and silence a NuGet warning for LibUsbDotNet. Standardize native signatures: use UIntPtr for read/ioctl counts/requests and change dirent struct naming for 64-bit. Windows interop: switch DllImport CharSet to Unicode, adjust DeviceIoControl overloads to use uint sizes and update SetupDi/GetClassDevs signature. HarmonyOS/OpenHarmony: use IntPtr for interface altsetting, add new DDK imports, reuse device mmap buffer instead of frequent allocations, and tighten probing logic/paths. Linux/OpenHarmony device code casts ioctl requests to UIntPtr and fixes partial-transfer handling. Rewrite LegacyUsbDevice to use proper Win32 ReadFile/WriteFile and DeviceIoControl semantics, implement ControlTransfer, add disposal/handle checks, and clean up trace/error handling throughout backends.

+++ FirmwareKit.Comm/Usb/Backend/macOS/MacOSUsbAPI.cs
@@ -70,13 +70,13 @@ internal static class MacOSUsbAPI
     [StructLayout(LayoutKind.Sequential)]
     public struct IOUSBFindInterfaceRequest
     {
-        public byte bInterfaceClass;
-        public byte bInterfaceSubClass;
-        public byte bInterfaceProtocol;
-        public byte bAlternateSetting;
+        public ushort bInterfaceClass;
+        public ushort bInterfaceSubClass;
+        public ushort bInterfaceProtocol;
+        public ushort bAlternateSetting;
     }
 
-    public const byte kIOUSBFindInterfaceDontCare = 0xFF;
+    public const ushort kIOUSBFindInterfaceDontCare = 0xFF;
 
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
     public delegate int USBDeviceOpenDelegate(IntPtr self);
@@ -123,6 +123,11 @@ internal static class MacOSUsbAPI
     public const int Offset_IUnknown_QueryInterface = 1;
     public const int Offset_IUnknown_Release = 3;
 
+    // These vtable offsets were determined through runtime probing on macOS.
+    // They may differ from the 0-based indices in Apple SDK header struct definitions
+    // (IOUSBDeviceInterface197 / IOUSBInterfaceInterface197) because the IOKit COM
+    // runtime may include undocumented reserved entries in the vtable.
+    // Do NOT modify these values based solely on header-file analysis.
     public const int Offset_DeviceRequest = 7;
     public const int Offset_USBDeviceOpen = 14;
     public const int Offset_USBDeviceClose = 15;
@@ -155,6 +160,11 @@ internal static class MacOSUsbAPI
     [UnmanagedFunctionPointer(CallingConvention.Cdecl)]
     public delegate int ClearPipeStallBothEndsDelegate(IntPtr self, byte pipeRef);
 
+    // These vtable offsets were determined through runtime probing on macOS.
+    // They may differ from the 0-based indices in Apple SDK header struct definitions
+    // (IOUSBInterfaceInterface197) because the IOKit COM runtime may include
+    // undocumented reserved entries in the vtable.
+    // Do NOT modify these values based solely on header-file analysis.
     public const int Offset_USBInterfaceOpen = 8;
     public const int Offset_USBInterfaceClose = 9;
     public const int Offset_GetNumEndpoints = 17;
